refactor(dashboard): split the pure helpers out of data_service.py (#1105 Phase 3 D5, develop-v2)
The 16 parsers/aggregators behind the DataService poll loop, plus WorkerLifecycle, move to
service/data_helpers.py; their 14 test classes move 1:1 into tests/service/test_data_helpers.py.
Pure move: every moved block is byte-identical to its original apart from one declared comment
edit and four declared test repoints.

The DataService class, its poll loop and the six poll-loop cadence/rate-limit constants that
interleave with the moved run all stay, per the cut map's binding ruling.
